+++ .github/workflows/ci.yml
@@ -7,18 +7,34 @@ on:
     branches: [master]
   schedule:
     - cron: '0 16 */15 * *'
+  workflow_dispatch:
+    inputs:
+      force_run:
+        description: 'Force workflow run'
+        required: true
+        type: choice
+        options: [yes, no]
+
+permissions:
+  actions: read
+  contents: read
+  statuses: write
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+env:
+  SRC_DIR: src/github.com/${{ github.repository }}
 
 jobs:
   Go:
     name: Go
     runs-on: ubuntu-latest
 
-    env:
-      SRC_DIR: src/github.com/${{ github.repository }}
-
     strategy:
       matrix:
-        go: [ '1.18.x', '1.19.x' ]
+        go: [ '1.19.x', '1.20.x' ]
 
     steps:
       - name: Set up Go
