+++ go.mod
@@ -2,6 +2,9 @@ module github.com/essentialkaos/uc
 
 go 1.18
 
-require github.com/essentialkaos/ek/v12 v12.60.1
+require (
+	github.com/essentialkaos/depsy v1.0.0
+	github.com/essentialkaos/ek/v12 v12.60.1
+)
 
 require golang.org/x/sys v0.6.0 // indirect
